fixing build

+++ src/MphRead.Native/Mods/Network/DedicatedServer.hpp
@@ -21,7 +21,7 @@ namespace System::Net
 
 namespace MphRead
 {
-    enum class GameMode : std::int32_t;
+    enum class GameMode : std::uint8_t;
 }
 
 namespace MphRead::Mods::Network
